GG-31194: Fix C++ Thin Tx memory leak (#1506)

+++ modules/platforms/cpp/odbc-test/src/odbc_test_suite.cpp
@@ -352,7 +352,7 @@ namespace ignite
                 for (size_t j = 0; j < valLen; ++j)
                 {
                     BOOST_TEST_INFO("Byte index: " << j);
-                    BOOST_CHECK_EQUAL(val[j], expected[j]);
+                    BOOST_CHECK_EQUAL(val[j], expected[(int32_t)j]);
                 }
             }
         }
